move: all the route handlers to `endpoints` directory
